#3 - Renamed project to 'nereides-javax'.
