+++ README.md
@@ -1,2 +1,2 @@
-# nereides-jdk
+# nereides-javax
 Object oriented JSON library wrapper for javax.json.
