+++ src/main/java/com/vzurauskas/nereides/javax/Checked.java
@@ -1,4 +1,4 @@
-package com.vzurauskas.nereides.jdk;
+package com.vzurauskas.nereides.javax;
 
 interface Checked<T> {
     T value() throws Exception;
